requirements.txt
Appium-Python-Client
selene~=2.0.0rc9
pytest~=8.3.4
selenium~=4.27.1